fixed encoder

+++ Node2_noLinux/GccApplication3/sam_pi_controller.c
@@ -1,8 +1,8 @@
 #include "sam_pi_controller.h"
 
-volatile uint8_t error_sum = 0;
+volatile uint16_t error_sum = 0;
 
-int16_t PI_controller(int16_t current_pos, int8_t target_pos) {
+int16_t PI_controller(int16_t current_pos, int16_t target_pos) {
 	int16_t error = target_pos - current_pos;
 	error_sum += error;
 	int16_t p = K_p * error;
